Move post items to post dir
