+++ src/components/post/postForm/index.ts
