+++ src/components/post/postItem/index.ts
