--- src/components/profilePicture/index.ts
@@ -1,1 +1,0 @@
-export { default } from './profilePicture.compnent';
